requirements.txt
Django==2.2.6
django-heroku==0.3.1
daphne==2.4.0
gunicorn==19.9.0
requests==2.22.0
channels==2.3.1
channels-redis==2.4.1
redis==3.3.11
social-auth-app-django==3.1.0
python-jose==3.1.0
python-dotenv==0.13.0
aiohttp==3.6.2
